requirements.txt
cuda-bindings==13.0.3
cuda-core==0.4.1
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cuda-toolkit==13.0.1
numpy==2.3.4
